organize project structure
re-organize some components into organisms. Create HomePageTemplate, modified ProfileTemplate to add Header and Footer
