--- yaqbot/src/components/molecules/NavMenu/NavMenu.module.css
@@ -1,10 +1,0 @@
-.navMenu {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.navItem {
-  padding: 1rem;
-  border: 1px solid #000;
-}
