+++ yaqbot/src/components/molecules/ProfileMenu/ProfileMenu.module.css
@@ -4,6 +4,7 @@
   position: absolute;
   top: 80px;
   right: 0;
+  z-index: 4;
 }
 
 .menu_option {
