+++ yaqbot/src/components/organisms/Header/Header.module.css
@@ -42,3 +42,9 @@
   padding: 0;
   margin: 0;
 }
+
+.nav_menu {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
