+++ yaqbot/src/components/organisms/ProfileDetails/ProfileDetails.tsx
@@ -13,7 +13,7 @@ interface DetailsBoardProps {
 const ProfileDetails: React.FC<DetailsBoardProps> = ({ bioDetails }) => {
   return (
     <Card className={styles.dashboard_card}>
-      <h1 className={styles.section_header}>Biografia del usuario</h1>
+      <h1 className={styles.section_header}>Perfil del Usuario</h1>
       <div className={styles.details_board}>
         {bioDetails.map((detail, index: number) => {
           return <BioDetail id={index} detailName={detail.label} description={detail.description} className={styles.detail} />;
