+++ yaqbot/src/components/organisms/ProfileSocials/ProfileSocials.module.css
@@ -21,6 +21,7 @@
   text-align: center;
   gap: 1rem;
   background-color: rgba(0, 0, 0, 0.2);
+  margin-top: 1rem;
 }
 
 .link {
