+++ yaqbot/src/components/pages/HomePage.tsx
@@ -1,0 +1,7 @@
+import HomeTemplate from '../templates/HomeTemplate/HomeTemplate';
+
+function HomePage() {
+  return <HomeTemplate />;
+}
+
+export default HomePage;
