+++ yaqbot/src/components/templates/HomeTemplate/HomeTemplate.module.css
@@ -1,11 +1,7 @@
-.homePage {
+.home {
   display: flex;
   flex-direction: column;
   justify-content: space-between;
   align-items: center;
   height: 100vh;
-}
-
-.aux {
-  background-color: rgba(0, 0, 0, 0.5);
 }
