+++ yaqbot/src/components/templates/ProfileTemplate/ProfileTemplate.module.css
@@ -1,9 +1,18 @@
+.page {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+}
+
 .profile {
   display: grid;
-  grid-template-columns: 3fr 7fr;
+  grid-template-columns: 3fr 6fr;
   grid-template-rows: auto auto;
   grid-template-areas:
     'a b'
     'c c';
   gap: 10px;
+  z-index: 1;
+  width: 95%;
 }
