+++ yaqbot/src/routes/routes.tsx
@@ -1,5 +1,5 @@
 import LoginPage from '../components/pages/LoginPage';
-import HomePage from '../components/pages/HomePage/HomePage';
+import HomePage from '../components/pages/HomePage';
 import SigninPage from '../components/pages/SigninPage';
 
 export type AppRoute = {
